# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice for storing and managing transactions",
        "title": "Transactions microservice",
        "contact": {
            "name": "Dias Galikhanov",
            "email": "diasgalikhanov@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8587",
    "basePath": "/api/v1",
    "paths": {
        "/transaction": {
            "get": {
                "description": "list transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "List Transactions",
                "operationId": "list-transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "create transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Create Transaction",
                "operationId": "create-transaction",
                "parameters": [
                    {
                        "description": "transaction info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transaction/{transaction_id}": {
            "get": {
                "description": "get transaction by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Get Transaction",
                "operationId": "get-transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction id",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "update transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Update Transaction",
                "operationId": "update-transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction id",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "transaction info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete transcation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Delete Transaction",
                "operationId": "delete-transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction id",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Transaction": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_amount": {
                    "type": "number"
                },
                "payment_time": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                }
            }
        }
    }
}